feat: add parametric fieldName and typeName on plugin options
With this changes a user can decide which fieldName to use and which typeName to use.
This commit comes with a new example folder `typename-fieldname` where this feature can be tested.

The primary motivation of this feature is because of the prismic query size limit (see [issue #15](https://github.com/birkir/gatsby-source-prismic-graphql/issues/15)).
With this changes a user can change the typeName to something smaller than "PRISMIC",
like "PRS"; this difference of characters, multiplied the times where on a query is
needed to match the fragments (`...on PRS_MySliceType` on every slice) can reduce the
query size.